status: report daemonset NamespacedName
Report in the object status, after succesful deployment
(condition=Available) the daemonset namespaced name.

This allows client to easily and safely (= we support and recommend this
flow) find which is the daemonset which manages RTEs.

This also allows cleaner internal flow, because previously we
relied on manifests to learne the namespaced name of the daemonset.
But this is wrong: we need use the observed state, not the desired
state.

Signed-off-by: Francesco Romani <fromani@redhat.com>