codegen: add tooling to autogenerate the clientset
We will need a RTE clientset in the upcoming e2e tests.
We can't just use the controller-runtime client because
it conflicts with the e2e framework from kube: both try to
register the` -kubeconfig` flag, leading to a startup panic.

We want to keep using the e2e framework from kube because
the testsuite inherited from RTE does so, and it's nice to be
consistent.

We want to import the e2e testsuite from RTE in order to
1. be able to run a sanity check the operand we deployed works as
   intended
2. be able to verify the RTE version we bundle (operator-as-operand)
   works as intended

Signed-off-by: Francesco Romani <fromani@redhat.com>